changed Pool::WorkerCountType back to unsigned int

+++ threadpool11/include/threadpool11/Pool.h
@@ -62,7 +62,7 @@ class Pool
 friend class Worker;
 
 public:
-	typedef size_t WorkerCountType;
+	typedef unsigned int WorkerCountType;
 
 private:
 	Pool(Pool&&);
